# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TTR ONE — ERP Platform API
  version: 0.1.0
  description: |
    REST API for **TTR ONE**, a multi-tenant SaaS ERP for Uzbekistan
    (Node/Fastify + PostgreSQL/Prisma backend, Nuxt 3 SPA frontend).

    ## Conventions
    - **Base URL**: all endpoints live under `/api/v1`.
    - **Auth**: JWT Bearer. Obtain an `accessToken` from `POST /auth/login`
      (15-minute TTL) and rotate it via `POST /auth/refresh`.
    - **Money**: all monetary amounts are integers in **minor units (tiyin)**,
      e.g. `1 000 000` = 10 000.00 UZS. Fields end in `Minor`.
    - **Quantities**: decimal strings, bounded to 6 decimal places.
    - **Multi-tenancy**: every record is scoped to the caller's tenant
      (derived from the token); there is no tenant id in the path.
    - **Errors**: uniform envelope `{ "error": { "code": "...", "message": "..." } }`.
    - **Pagination**: list endpoints accept `page`, `pageSize` (max 200),
      `sort`, `order` (`asc`/`desc`), `search`, and return a `meta` object
      `{ page, pageSize, total, totalPages }`.
  contact:
    name: TTR ONE
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
security:
  - apiKeyAuth: []
  - bearerAuth: []
...
paths:
...
  /catalog/products:
    get:
      tags: [catalog]
      summary: List products (paginated, searchable, sortable)
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
        - $ref: '#/components/parameters/sort'
        - $ref: '#/components/parameters/order'
        - $ref: '#/components/parameters/search'
      responses:
        '200':
          description: Products
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items: { $ref: '#/components/schemas/Product' }
                  meta: { $ref: '#/components/schemas/PageMeta' }
                  priceVisible: { type: boolean }
    post:
      tags: [catalog]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  product: { $ref: '#/components/schemas/Product' }
        '409': { $ref: '#/components/responses/Conflict' }
  /catalog/products/{id}:
    patch:
      tags: [catalog]
      summary: Update a product (SKU immutable)
      parameters: [{ $ref: '#/components/parameters/id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductInput' }
      responses:
        '200': { description: Updated }
        '404': { $ref: '#/components/responses/NotFound' }
  /catalog/products/{id}/barcode:
    post:
      tags: [catalog]
      summary: Generate a unique EAN-13 barcode for a product
      parameters: [{ $ref: '#/components/parameters/id' }]
      responses:
        '200': { description: Product with barcode }
  /catalog/products/by-barcode/{code}:
    get:
      tags: [catalog]
      summary: Look up a product by barcode (scan)
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Product }
        '404': { $ref: '#/components/responses/NotFound' }
  /catalog/units:
    get:
      tags: [catalog]
      summary: List units of measure
      responses: { '200': { description: Units } }
    post:
      tags: [catalog]
      summary: Create a unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, code]
              properties:
                name: { type: string }
                code: { type: string }
                precision: { type: integer, default: 0 }
      responses: { '201': { description: Created } }
  /catalog/categories:
    get:
      tags: [catalog]
      summary: List categories
      responses: { '200': { description: Categories } }
    post:
      tags: [catalog]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, code]
              properties:
                name: { type: string }
                code: { type: string }
                parentId: { type: string }
      responses: { '201': { description: Created } }

  # ==================== WAREHOUSE ====================
...
  /billing/subscribe:
    post:
      tags: [billing]
      summary: Create an invoice for a paid plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan]
              properties:
                plan: { type: string, enum: [starter, business] }
                method: { type: string, enum: [bank_transfer, card], default: bank_transfer }
      responses: { '201': { description: Invoice created } }
  /billing/pay:
    post:
      tags: [billing]
      summary: Pay an invoice by card (sandbox) and activate the plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [invoiceId]
              properties:
                invoiceId: { type: string }
                method: { type: string, enum: [card, mock, payme, click, stripe], default: card }
      responses: { '200': { description: Activated } }

  # ==================== ADMIN ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Access token from POST /auth/login, sent as `Authorization: Bearer <token>`.'
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 'Developer API key (ttr_live_…) created in the app under Управление → API-ключи. Recommended for server-to-server integrations. May also be sent as `Authorization: Bearer ttr_live_…`.'
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string }
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    pageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    sort:
      name: sort
      in: query
      schema: { type: string }
    order:
      name: order
      in: query
      schema: { type: string, enum: [asc, desc] }
    search:
      name: search
      in: query
      schema: { type: string }
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Unique constraint / duplicate
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string }
            details: { type: object, nullable: true }
    PageMeta:
      type: object
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
...
    Product:
      type: object
      properties:
        id: { type: string }
        sku: { type: string }
        name: { type: string }
        type: { type: string, enum: [stockable, service] }
        categoryId: { type: string, nullable: true }
        unitId: { type: string, nullable: true }
        priceMinor:
          type: integer
          nullable: true
          description: 'Price in minor units (tiyin); null when the caller lacks catalog.price'
        currency: { type: string }
        barcode: { type: string, nullable: true }
        tracking: { type: string, enum: [none, batch, serial], nullable: true }
    ProductInput:
      type: object
      required: [sku, name]
      properties:
        sku: { type: string }
        name: { type: string }
        type: { type: string, enum: [stockable, service], default: stockable }
        categoryId: { type: string }
        unitId: { type: string }
        priceMinor: { type: integer, default: 0, description: 'Minor units (tiyin)' }
        currency: { type: string, default: UZS }
        barcode: { type: string }
        tracking: { type: string, enum: [none, batch, serial] }
...